feat(foundation-demo): add backdrop animation states; separate backdrop animation from sidebar animations
add backdrop animation states; separate backdrop animation from sidebar animations
