+++ apps/foundation-demo/src/app/design/molecules/backdrop/animation/backdrop-animation-state.ts
@@ -1,0 +1,8 @@
+export const getAnimationState = (show : boolean) => {
+  if(show){
+    return "show";
+  }
+  else {
+    return "void";
+  }
+}
