+++ apps/foundation-demo/src/app/design/molecules/sidebar/animation/sidebar-animation-state.spec.ts
